separated the program's logic
