+++ zuck_small.blocks
@@ -1,6 +1,6 @@
-0 13 17 10 64395.01407 95076.276 71550.01563 71550
-1 14 17 10 64503.01407 148258.148 71670.01563 71670
-2 15 17 10 65016.00703 223559.0336 72240.00781 72240
+0 13 17 10 64395.01407 95076.276 0 71550
+1 14 17 10 64503.01407 148258.148 0 71670
+2 15 17 10 65016.00703 223559.0336 0 72240
 3 16 17 10 65124 372664.0976 72360 72360
 4 17 17 10 65124 531604.448 72360 72360
 5 18 17 10 65124 603718.6768 72360 72360
